new code refactor

+++ src/hive_viriatum/handlers/handler_dispatch.c
@@ -335,7 +335,7 @@ ERROR_CODE _sendResponseHandlerDispatch(struct HttpParser_t *httpParser) {
     SPRINTF(responseBuffer, 256, "HTTP/1.1 500 Internal Server Error\r\nServer: %s/%s (%s @ %s)\r\nConnection: Keep-Alive\r\nContent-Length: %lu\r\n\r\n%s", VIRIATUM_NAME, VIRIATUM_VERSION, VIRIATUM_PLATFORM_STRING, VIRIATUM_PLATFORM_CPU, (long unsigned int) sizeof(DISPATCH_ERROR_MESSAGE) - 1, DISPATCH_ERROR_MESSAGE);
 
     /* writes the response to the connection, registers for the appropriate callbacks */
-    writeConnection(connection, (unsigned char *) responseBuffer, (unsigned int) strlen(responseBuffer), _sendResponseCallbackHandlerDispatch, (void *) ((size_t) httpParser->flags));
+    writeConnection(connection, (unsigned char *) responseBuffer, (unsigned int) strlen(responseBuffer), _sendResponseCallbackHandlerDispatch, (void *) (size_t) httpParser->flags);
 
     /* raise no error */
     RAISE_NO_ERROR;
@@ -343,7 +343,7 @@ ERROR_CODE _sendResponseHandlerDispatch(struct HttpParser_t *httpParser) {
 
 ERROR_CODE _sendResponseCallbackHandlerDispatch(struct Connection_t *connection, struct Data_t *data, void *parameters) {
     /* retrieves the current http flags */
-    unsigned char flags = (unsigned char) parameters;
+    unsigned char flags = (unsigned char) (size_t) parameters;
 
     /* retrieves the underlying connection references in order to be
     able to operate over them, for unregister */
